[Refactor][Core] Split out header verification functions from main
